backend/dsc_api/requirements.txt
Django>=4.2
djangorestframework
gunicorn>=21.2
python-dotenv
pytest
pytest-django
django-cors-headers
pandas
scipy
matplotlib
plotly